編集対象ごとにファイル分割
